refactor: Organize example files
