+++ example/lib/main_development.dart
@@ -1,6 +1,6 @@
 import 'package:flavorbanner/flavorbanner.dart';
 import 'package:flutter/material.dart';
-import './lib/app.dart';
+import 'src/app.dart';
 
 void main() {
   FlavorConfig(
