+++ example/lib/main_production.dart
@@ -1,7 +1,7 @@
+import 'package:example/src/custom_values.dart';
 import 'package:flavorbanner/flavorbanner.dart';
 import 'package:flutter/material.dart';
-import './lib/app.dart';
-import './lib/custom_values.dart';
+import 'src/app.dart';
 
 void main() {
   FlavorConfig(
