+++ example/lib/main_staging.dart
@@ -1,13 +1,13 @@
 import 'package:flavorbanner/flavorbanner.dart';
 import 'package:flutter/material.dart';
-import './lib/app.dart';
+import 'src/app.dart';
 
 void main() {
   FlavorConfig(
-    flavor: Flavor.TEST,
-    color: Colors.orange,
+    flavor: Flavor.STAGING,
+    color: Colors.green,
     values: FlavorValues(
-      baseUrl: "https://test.com/",
+      baseUrl: "https://staging.com/",
       showBanner: true,
     ),
   );
