+++ example/lib/src/app.dart
